feat: 사용자 서비스 인터페이스 및 구현 클래스 추가
- 사용자 서비스 인터페이스 생성 및 메서드 정의
- 비즈니스 로직을 처리하는 AppUserService 클래스 추가
- 트랜잭션 경계를 설정하는 TxUserService 클래스 추가
